src/requirements.txt
beautifulsoup4==4.12.3
fastapi==0.111.0
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_openai==0.1.7
openai==1.30.1
pinecone==4.0.0
pydantic==2.7.1
python-dotenv==1.0.1
Requests==2.31.0
setuptools==69.5.1
uvicorn==0.29.0